add safe frames

+++ src/mod.h
@@ -9,7 +9,7 @@
 typedef struct _mt_mod {
     uint32_t ref_count;
     uint16_t _f_size, f_len;
-    size_t _size, len;
+    uint32_t _size, len;
     uint8_t* bytes;
     uint8_t* fns;
     // @TODO hash exports
